+++ src/Pages/RankPage/TopRankList.tsx
@@ -1,24 +1,11 @@
 import styled from '@emotion/styled';
-import { useMemo } from 'react';
 import { ChracterBox } from './ChracterBox';
 import { theme } from '@/styles/theme';
 import type { RankingUser } from './types';
+import { useTopRankList } from './hooks/useTopRankList';
 
 export const TopRankList = ({ topRankingUsers }: { topRankingUsers: RankingUser[] }) => {
-  const sortedByPoint = useMemo(
-    () => [...topRankingUsers].sort((a, b) => b.point - a.point),
-    [topRankingUsers],
-  );
-  const top3 = useMemo(() => sortedByPoint.slice(0, 3), [sortedByPoint]);
-  const arranged = useMemo(() => {
-    const [first, second, third] = top3;
-    return [second, first, third].filter(Boolean) as RankingUser[];
-  }, [top3]);
-
-  const rankLabels = useMemo(() => {
-    const labels = [2, 1, 3];
-    return labels.slice(0, arranged.length);
-  }, [arranged]);
+  const { arranged, rankLabels } = useTopRankList(topRankingUsers);
 
   return (
     <Wrapper>
